refactor: streamline resolver and service methods to use Either for error handling
- Refactored TransactionResolver to simplify dependency injection and method signatures.
- Updated UserResolver to handle user retrieval and balance calculations with Either.
- Enhanced UserService to validate user input using zod and return Either for error handling.
- Modified CategoryService and TransactionService to return Either for operations.
- Improved test utilities for mocking services and contexts.
- Adjusted tests to align with new Either-based error handling approach.

+++ src/models/category-summary.model.ts
@@ -1,13 +1,13 @@
-import { Field, ID, Int, ObjectType } from 'type-graphql'
+import { Field, Float, ID, Int, ObjectType } from 'type-graphql'
 @ObjectType()
 export class CategoriesSummaryModel {
   @Field(() => [CategoriesAggregatedModel])
   categories!: CategoriesAggregatedModel[]
 
-  @Field(() => Int)
+  @Field(() => Float)
   transactionCountByUser!: number
 
-  @Field(() => Int)
+  @Field(() => Float)
   categoryCount!: number
 
   @Field(() => CategoriesAggregatedModel)
@@ -28,7 +28,7 @@ export class CategoriesAggregatedModel {
   @Field(() => String)
   icon!: string
 
-  @Field(() => Int)
+  @Field(() => Float)
   totalAmount!: number
 
   @Field(() => Int)
